api_gateway_python/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.2.0
python-multipart==0.0.6
python-dotenv==1.0.1
httpx==0.26.0
redis==5.0.1
psycopg2-binary==2.9.9